+++ backend/app/schemas/finanzas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Literal
 
 from pydantic import BaseModel, Field
@@ -13,7 +14,7 @@ TipoPago = Literal[
 
 
 class OperacionPagoPayload(BaseModel):
-    monto: float = Field(gt=0)
+    monto: Decimal = Field(gt=0, max_digits=10, decimal_places=2)
     metodo: str = Field(default="Efectivo", max_length=100)
     motivo: str = Field(default="", max_length=1000)
     referencia: str = Field(default="", max_length=150)
